+++ framework/memory/src/core/type.ts
@@ -10,7 +10,7 @@ export interface VectorStoreSearchOptions {
 export interface Retriever<T> {
   get(id: string): Promise<VectorStoreDocument<T> | null>;
   insert(document: VectorStoreDocument<T>): Promise<void>;
-  delete(id: string): Promise<void>;
+  delete(idOrFilter: string | string[] | Record<string, any>): Promise<VectorStoreDocument<T>[]>;
   reset(): Promise<void>;
   update(document: VectorStoreDocument<T>): Promise<void>;
   list(k: number, options?: VectorStoreSearchOptions): Promise<VectorStoreDocument<T>[]>;
@@ -44,7 +44,7 @@ export interface MessageHistory {
 }
 
 export interface HistoryStore<T> {
-  addHistory(history: Omit<ActionHistory<T>, 'createdAt' | 'updatedAt'>): Promise<ActionHistory<T>>;
+  addHistory(...histories: Omit<ActionHistory<T>, 'createdAt' | 'updatedAt'>[]): Promise<ActionHistory<T>[]>;
   getHistory(memoryId: string): Promise<ActionHistory<T>[]>;
   addMessage(history: Omit<MessageHistory, 'createdAt' | 'updatedAt'>): Promise<MessageHistory>;
   getMessages(options: { filter: { [key: string]: any } }): Promise<MessageHistory[]>;